jean-memory-v3-local/requirements.txt
# Jean Memory V3 Local Service Dependencies

# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
python-multipart==0.0.20

# Memory Frameworks
mem0ai==0.1.30
faiss-cpu>=1.7.0

# Graph Database
py2neo==2021.2.4
neo4j==5.28.0

# AI/ML
sentence-transformers==3.3.1
openai==1.58.1
transformers==4.47.1
torch>=2.0.0

# Google ADK Dependencies
google-ai-generativelanguage>=0.5.0
google-cloud-aiplatform>=1.40.0
google-auth>=2.20.0

# Utilities
httpx==0.28.1
python-dotenv==1.0.1
aiofiles==24.1.0
pydantic-settings>=2.0.0
numpy>=1.24.0

# Optional Docker Support
docker>=6.0.0

# Development & Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-benchmark>=4.0.0
pytest-cov>=4.0.0
locust>=2.20.0
black==24.10.0
isort==5.13.2